+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [preloader] support Bearer token Authentication for non temporal logs.
 * [preloader] support end indexes
 * [CTFE] Add a /log.v3.json endpoint to help satisfy a requirement of the Chrome CT Log Policy by @robstradling in https://github.com/google/certificate-transparency-go/pull/1703
+* [CTFE] Short cache max-age when get-entries returns fewer entries than requested by @robstradling in https://github.com/google/certificate-transparency-go/pull/1707
 
 ### CTFE Storage Saving: Extra Data Issuance Chain Deduplication